requirements.txt
gradio
pandas
numpy
pillow
openpyxl